+++ src/components/ResearchField/hooks/useResearchField.js
@@ -1,6 +1,6 @@
 import { useState, useEffect, useCallback } from 'react';
 import { getResource } from 'services/backend/resources';
-import { getParentResearchFields, getStatementsBySubjectAndPredicate } from 'services/backend/statements';
+import { getStatementsBySubjectAndPredicate } from 'services/backend/statements';
 import { getResearchFieldsStats } from 'services/backend/stats';
 import { orderBy } from 'lodash';
 import { useParams } from 'react-router-dom';
@@ -8,7 +8,6 @@ import { PREDICATES } from 'constants/graphSettings';
 
 function useResearchField(initialVal = {}) {
     const [data, setData] = useState({ initialVal });
-    const [parentResearchFields, setParentResearchFields] = useState([]);
     const [subResearchFields, setSubResearchFields] = useState([]);
     const { researchFieldId } = useParams();
     const [isLoadingData, setIsLoadingData] = useState(true);
@@ -29,10 +28,6 @@ function useResearchField(initialVal = {}) {
                     setIsLoadingData(false);
                     setIsFailedLoadingData(true);
                 });
-
-            getParentResearchFields(rfId).then(result => {
-                setParentResearchFields(result.reverse());
-            });
 
             getStatementsBySubjectAndPredicate({ subjectId: rfId, predicateId: PREDICATES.HAS_SUB_RESEARCH_FIELD }).then(result => {
                 if (result.length > 0) {
@@ -58,6 +53,6 @@ function useResearchField(initialVal = {}) {
             loadResearchFieldData(researchFieldId);
         }
     }, [researchFieldId, loadResearchFieldData]);
-    return [data, parentResearchFields, subResearchFields, isLoadingData, isFailedLoadingData, loadResearchFieldData];
+    return [data, subResearchFields, isLoadingData, isFailedLoadingData, loadResearchFieldData];
 }
 export default useResearchField;