+++ src/components/ViewPaper/ProvenanceBox/ProvenanceBox.js
@@ -43,14 +43,16 @@ export default function ProvenanceBox(props) {
                                         <StyledItemProvenanceBox>
                                             <b style={{ textTransform: 'uppercase' }}>{props.observatoryInfo.name}</b>
                                             <br />
-                                            <Link to={reverse(ROUTES.ORGANIZATION, { id: props.observatoryInfo.organization.id })}>
-                                                <img
-                                                    style={{ marginTop: 8, marginBottom: 8, maxWidth: '80%', height: 'auto' }}
-                                                    className="mx-auto d-block"
-                                                    src={props.observatoryInfo.organization.logo}
-                                                    alt=""
-                                                />
-                                            </Link>
+                                            {props.observatoryInfo.organization && props.observatoryInfo.organization.id && (
+                                                <Link to={reverse(ROUTES.ORGANIZATION, { id: props.observatoryInfo.organization.id })}>
+                                                    <img
+                                                        style={{ marginTop: 8, marginBottom: 8, maxWidth: '80%', height: 'auto' }}
+                                                        className="mx-auto d-block"
+                                                        src={props.observatoryInfo.organization.logo}
+                                                        alt=""
+                                                    />
+                                                </Link>
+                                            )}
                                             {/* <p> */}
                                             {/* <Link to={reverse(ROUTES.ORGANIZATION, { id: props.observatoryInfo.organization.id })}> */}
                                             {/* {props.observatoryInfo.organization.name} */}