+++ src/pages/Comparison.js
@@ -10,6 +10,7 @@ import SelectProperties from 'components/Comparison/SelectProperties';
 import ValuePlugins from 'components/ValuePlugins/ValuePlugins';
 import AddContribution from 'components/Comparison/AddContribution/AddContribution';
 import ProvenanceBox from 'components/Comparison/ProvenanceBox/ProvenanceBox';
+import Breadcrumbs from 'components/Breadcrumbs/Breadcrumbs';
 import RelatedResources from 'components/Comparison/RelatedResources';
 import RelatedFigures from 'components/Comparison/RelatedFigures';
 import ExportCitation from 'components/Comparison/ExportCitation';
@@ -60,6 +61,7 @@ function Comparison(props) {
         hasNextVersions,
         createdBy,
         provenance,
+        researchField,
         setMetaData,
         setComparisonType,
         toggleProperty,
@@ -138,6 +140,8 @@ function Comparison(props) {
 
     return (
         <div>
+            <Breadcrumbs researchFieldId={researchField ? researchField.id : null} />
+
             <ContainerAnimated className="d-flex align-items-center">
                 <h1 className="h4 mt-4 mb-4 flex-grow-1">
                     Contribution comparison{' '}
@@ -163,6 +167,7 @@ function Comparison(props) {
                                     <DropdownItem onClick={handleFullWidth}>
                                         <span className="mr-2">{fullWidth ? 'Reduced width' : 'Full width'}</span>
                                     </DropdownItem>
+                                    <DropdownItem onClick={() => toggleTranspose(v => !v)}>Transpose table</DropdownItem>
                                     <DropdownItem divider />
                                     <DropdownItem header>View density</DropdownItem>
                                     <DropdownItem active={viewDensity === 'spacious'} onClick={() => handleViewDensity('spacious')}>
@@ -192,7 +197,6 @@ function Comparison(props) {
                                 <DropdownMenu right>
                                     <DropdownItem header>Customize</DropdownItem>
                                     <DropdownItem onClick={() => setShowPropertiesDialog(v => !v)}>Select properties</DropdownItem>
-                                    <DropdownItem onClick={() => toggleTranspose(v => !v)}>Transpose table</DropdownItem>
                                     <DropdownItem divider />
                                     <DropdownItem header>Export</DropdownItem>
                                     <DropdownItem onClick={() => setShowLatexDialog(v => !v)}>Export as LaTeX</DropdownItem>
@@ -237,7 +241,7 @@ function Comparison(props) {
                                     <DropdownItem
                                         onClick={e => {
                                             if (!props.user) {
-                                                props.openAuthDialog('signin', true);
+                                                props.openAuthDialog({ action: 'signin', signInRequired: true });
                                             } else {
                                                 setShowPublishDialog(v => !v);
                                             }
@@ -254,10 +258,14 @@ function Comparison(props) {
                                             </DropdownItem>
                                         </>
                                     )}
-                                    <DropdownItem divider />
-                                    <DropdownItem tag={NavLink} exact to={reverse(ROUTES.RESOURCE, { id: metaData?.id })}>
-                                        View resource
-                                    </DropdownItem>
+                                    {metaData?.id && (
+                                        <>
+                                            <DropdownItem divider />
+                                            <DropdownItem tag={NavLink} exact to={reverse(ROUTES.RESOURCE, { id: metaData.id })}>
+                                                View resource
+                                            </DropdownItem>
+                                        </>
+                                    )}
                                 </DropdownMenu>
                             </Dropdown>
                         </ButtonGroup>
@@ -445,7 +453,7 @@ function Comparison(props) {
 
             <ExportToLatex
                 data={matrixData}
-                contributions={contributions}
+                contributions={contributions.filter(c => c.active)}
                 properties={properties}
                 showDialog={showLatexDialog}
                 toggle={() => setShowLatexDialog(v => !v)}
@@ -478,7 +486,7 @@ const mapStateToProps = state => ({
 });
 
 const mapDispatchToProps = dispatch => ({
-    openAuthDialog: (action, signInRequired) => dispatch(openAuthDialog(action, signInRequired))
+    openAuthDialog: payload => dispatch(openAuthDialog(payload))
 });
 
 Comparison.propTypes = {