+++ src/services/backend/misc.js
@@ -3,10 +3,6 @@ import { submitGetRequest, submitPostRequest } from 'network';
 
 export const doisUrl = `${url}dois/`;
 
-export const saveFullPaper = (data, mergeIfExists = false) => {
-    return submitPostRequest(`${url}papers/?mergeIfExists=${mergeIfExists}`, { 'Content-Type': 'application/json' }, data);
-};
-
 export const getPaperByDOI = doi => {
     return submitGetRequest(`${url}widgets/?doi=${doi}`);
 };