+++ src/utils.js
@@ -322,9 +322,11 @@ export const generateRdfDataVocabularyFile = (data, contributions, properties, m
     //components
     const columns = [
         { id: 'Properties', title: 'Properties' },
-        ...contributions.map((contribution, index) => {
-            return contribution;
-        })
+        ...contributions
+            .filter(c => c.active)
+            .map((contribution, index) => {
+                return contribution;
+            })
     ];
     columns.forEach(function(column, index) {
         if (column.id === 'Properties') {
@@ -359,18 +361,20 @@ export const generateRdfDataVocabularyFile = (data, contributions, properties, m
             gds.add(new rdf.Triple(bno, cubens('dataSet'), ds));
             gds.add(new rdf.Triple(bno, dt['Properties'].toString(), new rdf.Literal(property.label.toString())));
             contributions.map((contribution, index2) => {
-                const cell = data[property.id][index2];
-                if (cell.length > 0) {
-                    cell.map(v => {
-                        if (v.type && v.type === 'resource') {
-                            gds.add(new rdf.Triple(bno, dt[contribution.id].toString(), orkgResource(`${v.resourceId}`)));
-                        } else {
-                            gds.add(new rdf.Triple(bno, dt[contribution.id].toString(), new rdf.Literal(`${v.label ? v.label : ''}`)));
-                        }
-                        return null;
-                    });
-                } else {
-                    gds.add(new rdf.Triple(bno, dt[contribution.id].toString(), new rdf.Literal('Empty')));
+                if (contribution.active) {
+                    const cell = data[property.id][index2];
+                    if (cell.length > 0) {
+                        cell.map(v => {
+                            if (v.type && v.type === 'resource') {
+                                gds.add(new rdf.Triple(bno, dt[contribution.id].toString(), orkgResource(`${v.resourceId}`)));
+                            } else {
+                                gds.add(new rdf.Triple(bno, dt[contribution.id].toString(), new rdf.Literal(`${v.label ? v.label : ''}`)));
+                            }
+                            return null;
+                        });
+                    } else {
+                        gds.add(new rdf.Triple(bno, dt[contribution.id].toString(), new rdf.Literal('Empty')));
+                    }
                 }
                 return null;
             });
@@ -690,6 +694,43 @@ function getOrder(paperStatements) {
         order = Infinity;
     }
     return order;
+}
+
+/**
+ * Parse resources statements and return a related figures objects
+ * @param {Array} resourceStatements
+ */
+export function getRelatedFiguresData(resourcesStatements) {
+    const _figures = resourcesStatements.map(resourceStatements => {
+        const imageStatement = resourceStatements.statements.find(statement => statement.predicate.id === PREDICATES.IMAGE);
+        const alt = resourceStatements.statements.length ? resourceStatements.statements[0]?.subject?.label : null;
+        return {
+            src: imageStatement ? imageStatement.object.label : '',
+            figureId: resourceStatements.id,
+            alt
+        };
+    });
+    return _figures;
+}
+
+/**
+ * Parse resources statements and return a related resources objects
+ * @param {Array} resourceStatements
+ */
+export function getRelatedResourcesData(resourcesStatements) {
+    const _resources = resourcesStatements.map(resourceStatements => {
+        const imageStatement = resourceStatements.statements.find(statement => statement.predicate.id === PREDICATES.IMAGE);
+        const urlStatement = resourceStatements.statements.find(statement => statement.predicate.id === PREDICATES.URL);
+        const descriptionStatement = resourceStatements.statements.find(statement => statement.predicate.id === PREDICATES.DESCRIPTION);
+        return {
+            url: urlStatement ? urlStatement.object.label : '',
+            image: imageStatement ? imageStatement.object.label : '',
+            id: resourceStatements.id,
+            title: resourceStatements.statements[0]?.subject?.label,
+            description: descriptionStatement ? descriptionStatement.object.label : ''
+        };
+    });
+    return _resources;
 }
 
 /**